+++ src/util/misc/textStyles.ts
@@ -31,6 +31,7 @@ export const hasStyleChanged = (
   prevStyle.fontWeight !== thisStyle.fontWeight ||
   prevStyle.fontStyle !== thisStyle.fontStyle ||
   prevStyle.textDecorationThickness !== thisStyle.textDecorationThickness ||
+  prevStyle.textDecorationColor !== thisStyle.textDecorationColor ||
   prevStyle.textBackgroundColor !== thisStyle.textBackgroundColor ||
   prevStyle.deltaY !== thisStyle.deltaY ||
   (forTextSpans &&